# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Sistema Jurídico API",
    "description": "Documentação oficial da API do LegalTech",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/ia/resumos/audiencia": {
      "post": {
        "tags": [
          "Inteligência Artificial"
        ],
        "summary": "Gerar resumo preparatório de audiência via IA a partir de documentos dos autos",
        "description": "Baixa os PDFs anexados no Google Drive, extrai o texto processual e gera um dossiê tático estruturado com fatos incontroversos, fatos controvertidos, riscos processuais, roteiro de perguntas e parâmetros de acordo",
        "operationId": "resumirParaAudiencia",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumoPecaDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Lista de identificadores de documentos inválida ou vazia",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroValidacaoDTO"
                }
              }
            }
          },
          "422": {
            "description": "Falha ao baixar documentos ou processar resumo pelo modelo de IA",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "403": {
            "description": "Acesso proibido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "200": {
            "description": "Resumo estruturado gerado com sucesso pela inteligência artificial",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResumoAudienciaEstruturadoDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErroPadraoDTO": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "Data e hora em que o erro ocorreu",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "description": "Código HTTP do erro",
            "format": "int32",
            "example": 400
          },
          "error": {
            "type": "string",
            "description": "Identificador ou classificação do erro",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "description": "Descrição detalhada da mensagem de erro",
            "example": "Dados inválidos fornecidos"
          }
        },
        "description": "Estrutura padrão de resposta para erros da API"
      },
      "CampoErroDTO": {
        "type": "object",
        "properties": {
          "campo": {
            "type": "string",
            "description": "Nome do campo que falhou na validação",
            "example": "email"
          },
          "mensagem": {
            "type": "string",
            "description": "Mensagem descrevendo o motivo da falha",
            "example": "E-mail inválido"
          }
        },
        "description": "Detalhe do erro de validação em um campo específico"
      },
      "ErroValidacaoDTO": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "Data e hora em que o erro ocorreu",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "description": "Código HTTP do erro",
            "format": "int32",
            "example": 400
          },
          "error": {
            "type": "string",
            "description": "Identificador ou classificação do erro",
            "example": "Validation Error"
          },
          "message": {
            "type": "string",
            "description": "Mensagem geral do erro de validação",
            "example": "Erro de validação nos campos informados."
          },
          "fieldErrors": {
            "type": "array",
            "description": "Lista detalhada dos erros encontrados por campo",
            "items": {
              "$ref": "#/components/schemas/CampoErroDTO"
            }
          }
        },
        "description": "Estrutura de resposta para erros de validação de campos"
      },
...
      "ResumoPecaDTO": {
        "required": [
          "documentosIds"
        ],
        "type": "object",
        "properties": {
          "documentosIds": {
            "type": "array",
            "description": "Lista de identificadores dos documentos (UUID do sistema ou ID no Google Drive) a serem lidos e sumarizados pela IA",
            "example": [
              "1PPOtwVobdFDcWgy52WS7rpy0pehRiJAZ",
              "a2d5926a-939e-4c7b-94ec-7cfa8e1b1234"
            ],
            "items": {
              "type": "string",
              "description": "Lista de identificadores dos documentos (UUID do sistema ou ID no Google Drive) a serem lidos e sumarizados pela IA",
              "example": "[\"1PPOtwVobdFDcWgy52WS7rpy0pehRiJAZ\",\"a2d5926a-939e-4c7b-94ec-7cfa8e1b1234\"]"
            }
          }
        },
        "description": "Payload para requisição de resumo preparatório de audiência a partir de documentos dos autos"
      },
      "ResumoAudienciaEstruturadoDTO": {
        "type": "object",
        "properties": {
          "fatosIncontroversos": {
            "type": "array",
            "description": "Fatos já pacificados, incontroversos ou admitidos nos autos",
            "example": [
              "Vínculo empregatício formalizado no período de 01/02/2022 a 15/03/2024",
              "Demissão imotivada"
            ],
            "items": {
              "type": "string",
              "description": "Fatos já pacificados, incontroversos ou admitidos nos autos",
              "example": "[\"Vínculo empregatício formalizado no período de 01/02/2022 a 15/03/2024\",\"Demissão imotivada\"]"
            }
          },
          "fatosControvertidos": {
            "type": "array",
            "description": "Pontos de conflito fático que dependem de instrução probatória ou prova oral",
            "example": [
              "Cumprimento de horas extras habituais além da jornada contratual",
              "Existência de dano moral decorrente de assédio"
            ],
            "items": {
              "type": "string",
              "description": "Pontos de conflito fático que dependem de instrução probatória ou prova oral",
              "example": "[\"Cumprimento de horas extras habituais além da jornada contratual\",\"Existência de dano moral decorrente de assédio\"]"
            }
          },
          "riscosProcessuais": {
            "type": "array",
            "description": "Preliminares processuais, fragilidades da tese e riscos de sucumbência",
            "example": [
              "Risco de prescrição bienal ou quinquenal",
              "Fragilidade na prova documental de jornada"
            ],
            "items": {
              "type": "string",
              "description": "Preliminares processuais, fragilidades da tese e riscos de sucumbência",
              "example": "[\"Risco de prescrição bienal ou quinquenal\",\"Fragilidade na prova documental de jornada\"]"
            }
          },
          "roteiroPerguntas": {
            "type": "array",
            "description": "Roteiro de perguntas sugeridas para inquirição de testemunhas ou depoimento da parte contrária",
            "example": [
              "Indagar a testemunha se havia controle formal de ponto e se as horas extras eram registradas fielmente"
            ],
            "items": {
              "type": "string",
              "description": "Roteiro de perguntas sugeridas para inquirição de testemunhas ou depoimento da parte contrária",
              "example": "[\"Indagar a testemunha se havia controle formal de ponto e se as horas extras eram registradas fielmente\"]"
            }
          },
          "parametrosAcordo": {
            "type": "string",
            "description": "Parâmetros, diretrizes e estimativas táticas para formulação de propostas de acordo",
            "example": "Recomenda-se proposta conciliatória de até R$ 8.000,00 parcelado em 3x, considerando o risco probatório das horas extras"
          }
        },
        "description": "Dossiê estruturado preparatório para audiência gerado por Inteligência Artificial"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}